refactor(autostore): use AutoSaveStartEvent for saving settings
