+++ features/autostore/src/main/java/net/okocraft/box/feature/autostore/AutoStoreFeature.java
@@ -5,18 +5,14 @@ import net.okocraft.box.api.feature.AbstractBoxFeature;
 import net.okocraft.box.api.feature.Reloadable;
 import net.okocraft.box.feature.autostore.command.AutoStoreCommand;
 import net.okocraft.box.feature.autostore.gui.AutoStoreClickMode;
+import net.okocraft.box.feature.autostore.listener.AutoSaveListener;
 import net.okocraft.box.feature.autostore.listener.BoxPlayerListener;
 import net.okocraft.box.feature.autostore.listener.ItemListener;
 import net.okocraft.box.feature.autostore.message.AutoStoreMessage;
 import net.okocraft.box.feature.autostore.model.container.AutoStoreSettingContainer;
-import net.okocraft.box.feature.autostore.task.AutoStoreSettingSaveTask;
 import net.okocraft.box.feature.gui.api.mode.ClickModeRegistry;
 import org.bukkit.command.CommandSender;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 
 public class AutoStoreFeature extends AbstractBoxFeature implements Reloadable {
 
@@ -27,13 +23,11 @@ public class AutoStoreFeature extends AbstractBoxFeature implements Reloadable {
     }
 
     private final BoxPlayerListener boxPlayerListener = new BoxPlayerListener();
+    private final AutoSaveListener autoSaveListener = new AutoSaveListener();
     private final ItemListener itemListener = new ItemListener();
-    private final AutoStoreSettingSaveTask autoSaveTask = new AutoStoreSettingSaveTask();
 
     private final AutoStoreCommand autoStoreCommand = new AutoStoreCommand();
     private final AutoStoreClickMode autoStoreClickMode = new AutoStoreClickMode();
-
-    private ScheduledExecutorService scheduler;
 
     public AutoStoreFeature() {
         super("autostore");
@@ -42,12 +36,11 @@ public class AutoStoreFeature extends AbstractBoxFeature implements Reloadable {
     @Override
     public void enable() {
         CONTAINER.loadAll();
-        boxPlayerListener.register(getListenerKey());
-        itemListener.register();
 
-        scheduler = Executors.newSingleThreadScheduledExecutor();
-        scheduler.scheduleAtFixedRate(autoSaveTask, 10, 10, TimeUnit.MINUTES);
-        autoSaveTask.registerListener(getListenerKey());
+        boxPlayerListener.register(getListenerKey());
+        autoSaveListener.register(getListenerKey());
+
+        itemListener.register();
 
         BoxProvider.get().getBoxCommand().getSubCommandHolder().register(autoStoreCommand);
         ClickModeRegistry.register(autoStoreClickMode);
@@ -59,12 +52,9 @@ public class AutoStoreFeature extends AbstractBoxFeature implements Reloadable {
         ClickModeRegistry.unregister(autoStoreClickMode);
 
         itemListener.unregister();
-        boxPlayerListener.unregister();
 
-        if (scheduler != null) {
-            scheduler.shutdownNow();
-            autoSaveTask.unregisterListener(getListenerKey());
-        }
+        autoSaveListener.unregister();
+        boxPlayerListener.unregister();
 
         CONTAINER.unloadAll();
     }
