+++ features/autostore/src/main/java/net/okocraft/box/feature/autostore/listener/AutoSaveListener.java
@@ -1,7 +1,8 @@
-package net.okocraft.box.feature.autostore.task;
+package net.okocraft.box.feature.autostore.listener;
 
 import com.github.siroshun09.event4j.handlerlist.Key;
 import net.okocraft.box.api.BoxProvider;
+import net.okocraft.box.api.event.general.AutoSaveStartEvent;
 import net.okocraft.box.feature.autostore.AutoStoreFeature;
 import net.okocraft.box.feature.autostore.event.AutoStoreSettingChangeEvent;
 import net.okocraft.box.feature.autostore.model.setting.AutoStoreSetting;
@@ -12,12 +13,31 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 
-public class AutoStoreSettingSaveTask implements Runnable {
+public class AutoSaveListener {
 
     private final List<AutoStoreSetting> modifiedSettings = new ArrayList<>();
 
-    @Override
-    public void run() {
+    private Key listenerKey;
+
+    public void register(@NotNull Key listenerKey) {
+        this.listenerKey = listenerKey;
+
+        var eventBus = BoxProvider.get().getEventBus();
+
+        eventBus.getHandlerList(AutoStoreSettingChangeEvent.class)
+                .subscribe(listenerKey, event -> modifiedSettings.add(event.getSetting()));
+        eventBus.getHandlerList(AutoSaveStartEvent.class).subscribe(listenerKey, this::saveModifiedSettings);
+    }
+
+    public void unregister() {
+        if (listenerKey != null) {
+            var eventBus = BoxProvider.get().getEventBus();
+            eventBus.getHandlerList(AutoStoreSettingChangeEvent.class).unsubscribeAll(listenerKey);
+            eventBus.getHandlerList(AutoSaveStartEvent.class).unsubscribeAll(listenerKey);
+        }
+    }
+
+    private void saveModifiedSettings(@NotNull AutoSaveStartEvent event) {
         var copied = List.copyOf(modifiedSettings);
 
         modifiedSettings.clear();
@@ -30,26 +50,11 @@ public class AutoStoreSettingSaveTask implements Runnable {
                     task.exceptionallyAsync(throwable -> {
                         BoxProvider.get().getLogger().log(
                                 Level.SEVERE,
-                                "Could not unload autostore setting (" + setting.getUuid() + ")",
+                                "Could not save autostore setting (" + setting.getUuid() + ")",
                                 throwable
                         );
-
                         return null;
                     });
                 });
-    }
-
-    public void registerListener(@NotNull Key listenerKey) {
-        BoxProvider.get()
-                .getEventBus()
-                .getHandlerList(AutoStoreSettingChangeEvent.class)
-                .subscribe(listenerKey, event -> modifiedSettings.add(event.getSetting()));
-    }
-
-    public void unregisterListener(@NotNull Key listenerKey) {
-        BoxProvider.get()
-                .getEventBus()
-                .getHandlerList(AutoStoreSettingChangeEvent.class)
-                .unsubscribeAll(listenerKey);
     }
 }
